refactor(Tests): reorder class structure
