--- tests/Feature/ExampleTest.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace VOSTPT\Tests\Feature;
-
-use VOSTPT\Tests\TestCase;
-
-class ExampleTest extends TestCase
-{
-    //
-}
