+++ tests/Integration/CreatesApplication.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace VOSTPT\Tests;
+namespace VOSTPT\Tests\Integration;
 
 use Illuminate\Contracts\Console\Kernel;
 
@@ -15,7 +15,7 @@ trait CreatesApplication
      */
     public function createApplication()
     {
-        $app = require __DIR__.'/../bootstrap/app.php';
+        $app = require __DIR__.'/../../bootstrap/app.php';
 
         $app->make(Kernel::class)->bootstrap();
