--- tests/TestCase.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace VOSTPT\Tests;
-
-use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
-
-abstract class TestCase extends BaseTestCase
-{
-    use CreatesApplication;
-}
