--- tests/Unit/ExampleTest.php
@@ -1,11 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace VOSTPT\Tests\Unit;
-
-use VOSTPT\Tests\TestCase;
-
-class ExampleTest extends TestCase
-{
-    //
-}
