+++ tests/Unit/TestCase.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace VOSTPT\Tests;
+
+abstract class TestCase extends \PHPUnit\Framework\TestCase
+{
+    //
+}
